18-1-02
增加超管展示通道搜索条件
增加接口传单日志
增加展示通道维护状态字段，更新状态

+++ weizu-parent/weizu-web-modules/weizu-channel/src/main/java/com/weizu/flowsys/web/activity/dao/IChannelForShowDao.java
@@ -3,6 +3,6 @@ package com.weizu.flowsys.web.activity.dao;
 import com.weizu.flowsys.core.dao.Dao;
 import com.weizu.flowsys.web.activity.pojo.ChannelForShowPo;
 
-public interface IChannelForShowDao extends Dao<ChannelForShowPo, Integer> {
+public interface IChannelForShowDao extends Dao<ChannelForShowPo, Long> {
 
 }
