+++ weizu-parent/weizu-web-modules/weizu-channel/src/main/java/com/weizu/flowsys/web/activity/dao/impl/ChannelForShowDaoImpl.java
@@ -21,7 +21,7 @@ import com.weizu.flowsys.web.activity.pojo.ChannelForShowPo;
  * @version 1.0
  */
 @Repository(value="channelForShowDao")
-public class ChannelForShowDaoImpl extends DaoImpl<ChannelForShowPo, Integer> implements
+public class ChannelForShowDaoImpl extends DaoImpl<ChannelForShowPo, Long> implements
 		IChannelForShowDao {
